Backend/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
Django==4.1.2
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
mysqlclient==2.1.1
promise==2.3
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.4
graphene-file-upload==1.3.0
django-on-heroku==1.1.2
gunicorn==20.1.0
django-cors-headers==3.13.0
boto3==1.24.94
cryptography==38.0.1
pyjwt==2.6.0

